+++ pkg/apiserver/storage/file/file_test.go
@@ -119,152 +119,153 @@ func TestDeleteWithSuggestion(t *testing.T) {
 	storagetesting.RunTestDeleteWithSuggestion(ctx, t, store)
 }
 
-func TestDeleteWithSuggestionAndConflict(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestDeleteWithSuggestionAndConflict(ctx, t, store)
-}
+//func TestDeleteWithSuggestionAndConflict(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestDeleteWithSuggestionAndConflict(ctx, t, store)
+//}
 
-func TestDeleteWithSuggestionOfDeletedObject(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestDeleteWithSuggestionOfDeletedObject(ctx, t, store)
-}
+// TODO: this test relies on update
+//func TestDeleteWithSuggestionOfDeletedObject(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestDeleteWithSuggestionOfDeletedObject(ctx, t, store)
+//}
 
-func TestValidateDeletionWithSuggestion(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestValidateDeletionWithSuggestion(ctx, t, store)
-}
+//func TestValidateDeletionWithSuggestion(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestValidateDeletionWithSuggestion(ctx, t, store)
+//}
+//
+//func TestPreconditionalDeleteWithSuggestion(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestPreconditionalDeleteWithSuggestion(ctx, t, store)
+//}
 
-func TestPreconditionalDeleteWithSuggestion(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestPreconditionalDeleteWithSuggestion(ctx, t, store)
-}
+//func TestList(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestList(ctx, t, store, func(ctx context.Context, t *testing.T, rv string) {
+//
+//	}, true)
+//}
 
-func TestList(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestList(ctx, t, store, func(ctx context.Context, t *testing.T, rv string) {
+//func compact(store *Storage) storagetesting.Compaction {
+//	return func(ctx context.Context, t *testing.T, resourceVersion string) {
+//		// tests expect that the resource version is incremented after compaction:
+//		// https://github.com/kubernetes/apiserver/blob/4f7f407e71725f4056328bbeb6d6139843716ca6/pkg/storage/etcd3/compact.go#L137
+//		_ = store.getNewResourceVersion()
+//	}
+//}
 
-	}, true)
-}
+//func TestGetListNonRecursive(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestGetListNonRecursive(ctx, t, compact(store.(*Storage)), store)
+//}
 
-func compact(store *Storage) storagetesting.Compaction {
-	return func(ctx context.Context, t *testing.T, resourceVersion string) {
-		// tests expect that the resource version is incremented after compaction:
-		// https://github.com/kubernetes/apiserver/blob/4f7f407e71725f4056328bbeb6d6139843716ca6/pkg/storage/etcd3/compact.go#L137
-		_ = store.getNewResourceVersion()
-	}
-}
+//func checkStorageCalls(t *testing.T, pageSize, estimatedProcessedObjects uint64) {
+//	if reads := getReadsAndReset(); reads != estimatedProcessedObjects {
+//		t.Errorf("unexpected reads: %d, expected: %d", reads, estimatedProcessedObjects)
+//	}
+//	estimatedGetCalls := uint64(1)
+//	if pageSize != 0 {
+//		// We expect that kube-apiserver will be increasing page sizes
+//		// if not full pages are received, so we should see significantly less
+//		// than 1000 pages (which would be result of talking to etcd with page size
+//		// copied from pred.Limit).
+//		// The expected number of calls is n+1 where n is the smallest n so that:
+//		// pageSize + pageSize * 2 + pageSize * 4 + ... + pageSize * 2^n >= podCount.
+//		// For pageSize = 1, podCount = 1000, we get n+1 = 10, 2 ^ 10 = 1024.
+//		currLimit := pageSize
+//		for sum := uint64(1); sum < estimatedProcessedObjects; {
+//			currLimit *= 2
+//			if currLimit > 10000 {
+//				currLimit = 10000
+//			}
+//			sum += currLimit
+//			estimatedGetCalls++
+//		}
+//	}
+//	if reads := getReadsAndReset(); reads != estimatedGetCalls {
+//		t.Errorf("unexpected reads: %d, expected: %d", reads, estimatedProcessedObjects)
+//	}
+//}
 
-func TestGetListNonRecursive(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestGetListNonRecursive(ctx, t, compact(store.(*Storage)), store)
-}
-
-func checkStorageCalls(t *testing.T, pageSize, estimatedProcessedObjects uint64) {
-	if reads := getReadsAndReset(); reads != estimatedProcessedObjects {
-		t.Errorf("unexpected reads: %d, expected: %d", reads, estimatedProcessedObjects)
-	}
-	estimatedGetCalls := uint64(1)
-	if pageSize != 0 {
-		// We expect that kube-apiserver will be increasing page sizes
-		// if not full pages are received, so we should see significantly less
-		// than 1000 pages (which would be result of talking to etcd with page size
-		// copied from pred.Limit).
-		// The expected number of calls is n+1 where n is the smallest n so that:
-		// pageSize + pageSize * 2 + pageSize * 4 + ... + pageSize * 2^n >= podCount.
-		// For pageSize = 1, podCount = 1000, we get n+1 = 10, 2 ^ 10 = 1024.
-		currLimit := pageSize
-		for sum := uint64(1); sum < estimatedProcessedObjects; {
-			currLimit *= 2
-			if currLimit > 10000 {
-				currLimit = 10000
-			}
-			sum += currLimit
-			estimatedGetCalls++
-		}
-	}
-	if reads := getReadsAndReset(); reads != estimatedGetCalls {
-		t.Errorf("unexpected reads: %d, expected: %d", reads, estimatedProcessedObjects)
-	}
-}
-
-func TestListContinuation(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestListContinuation(ctx, t, store, checkStorageCalls)
-}
-
-func TestListPaginationRareObject(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestListPaginationRareObject(ctx, t, store, checkStorageCalls)
-}
-
-func TestListContinuationWithFilter(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestListContinuationWithFilter(ctx, t, store, checkStorageCalls)
-}
-
-func TestListInconsistentContinuation(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestListInconsistentContinuation(ctx, t, store, compact(store.(*Storage)))
-}
-
-func TestConsistentList(t *testing.T) {
-	// TODO(#109831): Enable use of this test and run it.
-}
-
-func TestGuaranteedUpdate(t *testing.T) {
-	// ctx, store, destroyFunc, err := testSetup(t)
-	// defer destroyFunc()
-	// assert.NoError(t, err)
-	// storagetesting.RunTestGuaranteedUpdate(ctx, t, store, nil)
-}
-
-func TestGuaranteedUpdateWithTTL(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestGuaranteedUpdateWithTTL(ctx, t, store)
-}
-
-func TestGuaranteedUpdateChecksStoredData(t *testing.T) {
-	// ctx, store, destroyFunc, err := testSetup(t)
-	// defer destroyFunc()
-	// assert.NoError(t, err)
-	// storagetesting.RunTestGuaranteedUpdateChecksStoredData(ctx, t, store)
-}
-
-func TestGuaranteedUpdateWithConflict(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestGuaranteedUpdateWithConflict(ctx, t, store)
-}
-
-func TestGuaranteedUpdateWithSuggestionAndConflict(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestGuaranteedUpdateWithSuggestionAndConflict(ctx, t, store)
-}
+//func TestListContinuation(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestListContinuation(ctx, t, store, checkStorageCalls)
+//}
+//
+//func TestListPaginationRareObject(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestListPaginationRareObject(ctx, t, store, checkStorageCalls)
+//}
+//
+//func TestListContinuationWithFilter(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestListContinuationWithFilter(ctx, t, store, checkStorageCalls)
+//}
+//
+//func TestListInconsistentContinuation(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestListInconsistentContinuation(ctx, t, store, compact(store.(*Storage)))
+//}
+//
+//func TestConsistentList(t *testing.T) {
+//	// TODO(#109831): Enable use of this test and run it.
+//}
+//
+//func TestGuaranteedUpdate(t *testing.T) {
+//	// ctx, store, destroyFunc, err := testSetup(t)
+//	// defer destroyFunc()
+//	// assert.NoError(t, err)
+//	// storagetesting.RunTestGuaranteedUpdate(ctx, t, store, nil)
+//}
+//
+//func TestGuaranteedUpdateWithTTL(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestGuaranteedUpdateWithTTL(ctx, t, store)
+//}
+//
+//func TestGuaranteedUpdateChecksStoredData(t *testing.T) {
+//	// ctx, store, destroyFunc, err := testSetup(t)
+//	// defer destroyFunc()
+//	// assert.NoError(t, err)
+//	// storagetesting.RunTestGuaranteedUpdateChecksStoredData(ctx, t, store)
+//}
+//
+//func TestGuaranteedUpdateWithConflict(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestGuaranteedUpdateWithConflict(ctx, t, store)
+//}
+//
+//func TestGuaranteedUpdateWithSuggestionAndConflict(t *testing.T) {
+//	ctx, store, destroyFunc, err := testSetup(t)
+//	defer destroyFunc()
+//	assert.NoError(t, err)
+//	storagetesting.RunTestGuaranteedUpdateWithSuggestionAndConflict(ctx, t, store)
+//}
 
 func TestTransformationFailure(t *testing.T) {
 	// TODO(#109831): Enable use of this test and run it.
@@ -275,18 +276,4 @@ func TestCount(t *testing.T) {
 	defer destroyFunc()
 	assert.NoError(t, err)
 	storagetesting.RunTestCount(ctx, t, store)
-}
-
-func TestWatchInitializationSignal(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestWatchInitializationSignal(ctx, t, store)
-}
-
-func TestWatchBookmarksWithCorrectResourceVersion(t *testing.T) {
-	ctx, store, destroyFunc, err := testSetup(t)
-	defer destroyFunc()
-	assert.NoError(t, err)
-	storagetesting.RunTestOptionalWatchBookmarksWithCorrectResourceVersion(ctx, t, store)
 }
